Add a module to publish broker events
The Module will be used to publish notifications to a medium(Ex:Message Broker) when an event is triggered in the message broker. It is needed to identify events when they are triggered. The module has EventService.class to register the service, Events.class to load needed Event Publisher, EventConfiguration.class to load the necessary configurations for the module and other supporting classes.